Qutip lab
Wigner function of an odd cat state. (left) Wigner function of an odd state compressed in the X direction.(center) Wigner function of an odd state compressed in the P direction. The fidelity and Wigner negativity in the P compression direction are higher, so it is more conducive to the protection of quantum states. This can be explained by the compression in the P direction making it more similar in frequency to the background, so it is less likely to decouple after passing through the loss channel, thus maintaining its Wigner negativity, that is, quantum properties, and having higher fidelity.